Add files via upload
Buffer overflow risk status is dominated by remote network penetration risk attempts to gain partial or complete control over an anonymous Internet user. If the buffer can 
effectively eliminate the risk of overflow, it can eliminate a large number of its most serious security threats. This article provides an in-depth survey of the different types of buffer overflow risk and attack. It also surveys the various safety measures required to reduce the risk of buffer overflows. Also, the methods presented in this article have the effect of accurately predicting the risk of software buffer overflow in C / C ++ and Java programs. By experimenting with techniques that can eliminate the problem of buffer overflow, it is possible to get rid of those safety hazards.